chore: reorganizing string folder (#42)
