+++ __tests__/string/formatters.test.ts
@@ -1,4 +1,4 @@
-import { capitalizeFirstLetter } from '../../../src/string'
+import { capitalizeFirstLetter } from '../../src/string'
 
 describe('capitalizeFirstLetter', () => {
   const data = [{ input: 'teste', output: 'Teste' }, { input: 'TESTE', output: 'TESTE' }, { input: 'alan Reno Neves', output: 'Alan Reno Neves' }]
